Refactoring and improving of get, create and delete CLI commands.

+++ cli/stackn/stackn.py
@@ -19,92 +19,7 @@ def _check_status(r, error_msg="Failed"):
     else:
         return True
 
-
-def get_endpoints(studio_url):
-    # These endpoints are related to the API available in Studio under: stackn/components/studio/api/
-    endpoints = dict()
-    if (not 'http://' in studio_url) and (not 'https://' in studio_url):
-        studio_url = 'https://'+studio_url
-    base = studio_url.strip('/')+'/api'
-    endpoints['models'] = base+'/projects/{}/models/'
-    endpoints['resources'] = base+'/projects/{}/resources/'
-    endpoints['appinstances'] = base+'/projects/{}/appinstances/'
-    endpoints['objecttypes'] = base+'/projects/{}/objecttype'
-    endpoints['members'] = base+'/projects/{}/members/'
-    endpoints['flavors'] = base+'/projects/{}/flavors/'
-    endpoints['environments'] = base+'/projects/{}/environments/'
-    endpoints['s3'] = base+'/projects/{}/s3/'
-    endpoints['mlflow'] = base+'/projects/{}/mlflow/'
-    endpoints['project_del'] = base+'/projects/{}'
-    endpoints['releasenames'] = base+'/projects/{}/releasenames/'
-    endpoints['projects'] = base+'/projects/'
-    endpoints['project_templates'] = base+'/projecttemplates/'
-    endpoints['admin'] = dict()
-    endpoints['admin']['apps'] = base+'/apps/'
-    return endpoints
-
-def get_auth_header(conf):
-    conf, status = stackn.auth.get_token(conf)
-    if not status:
-        return False, False
-    auth_header = {"Authorization": "Token {}".format(conf['STACKN_ACCESS_TOKEN'])}
-    return auth_header, conf
-
-def get_remote(inp_conf):
-    
-    conf, status = stackn.auth.get_config(inp_conf)
-
-    if not status:
-        return False
-
-    keys = stackn.auth._get_remote(conf)
-    
-    if not keys:
-        return False
-
-    return keys
-
-def get_current(secure):
-
-    res = {'STACKN_URL': False, 'STACKN_PROJECT': False}
-    conf, status = stackn.auth.get_config({'STACKN_SECURE': secure})
-
-    if not status:
-        print("Failed to get current STACKn configuration file.")
-    else:
-        if 'STACKN_URL' in conf:
-            res['STACKN_URL'] = conf['STACKN_URL']
-        if 'STACKN_PROJECT' in conf:
-            res['STACKN_PROJECT'] = conf['STACKN_PROJECT']
-    
-    return res
-
-def get_projects(conf={}, params=[], auth_header=[]):
-    conf, status = stackn.auth.get_config(conf)
-
-    if not status:
-        print('Cannot list projects.')
-        return False
-
-    auth_header, conf = get_auth_header(conf)
-
-    if not auth_header:
-        return False
-
-    endpoints = get_endpoints(conf['STACKN_URL'])
-    url = endpoints['projects']
-    if params:
-        r = requests.get(url, headers=auth_header, params=params, verify=conf['STACKN_SECURE'])
-    else:
-        r = requests.get(url, headers=auth_header, verify=conf['STACKN_SECURE'])
-    if r:
-        projects = json.loads(r.content)
-        return projects
-    else:
-        print("Fetching projects failed.")
-        print('Returned status code: {}'.format(r.status_code))
-        print('Reason: {}'.format(r.reason))
-        return None
+# Sort of utils functions
 
 def call_admin_endpoint(name, conf={}, params=[]):
     
@@ -135,7 +50,9 @@ def call_admin_endpoint(name, conf={}, params=[]):
         print('Reason: {}'.format(r.reason))
         return None
 
+
 def call_project_endpoint(name, conf={}, params=[]):
+
     conf, status = stackn.auth.get_config(conf)
 
     if not status:
@@ -159,10 +76,12 @@ def call_project_endpoint(name, conf={}, params=[]):
         print('Found several matching projects.')
         return
     if not project:
-        print("Project {} not found.".format(conf['STACKN_PROJECT']))
+        print("Project \'{}\' not found.".format(conf['STACKN_PROJECT']))
         return False
+    
     project = project[0]
     url = endpoints[name].format(project['id'])
+
     if params:
         r = requests.get(url, headers=auth_header, params=params, verify=conf['STACKN_SECURE'])
     else:
@@ -176,29 +95,125 @@ def call_project_endpoint(name, conf={}, params=[]):
         print('Reason: {}'.format(r.reason))
         return None
 
+
 def setup_project_endpoint_call(conf, endpoint_type):
+
     conf, status = stackn.auth.get_config(conf, required=['STACKN_URL'])
 
     if not status:
-        print("Missing required input (studio URL, project name).")
-        return False
+        print("Failed to get current STACKn configuration file.")
+        return False, False, False
 
     auth_header, conf = get_auth_header(conf)
     if not auth_header:
-        return False
+        return False, False, False
 
-    project = get_projects(conf=conf['STACKN_URL'], params={'name': conf['STACKN_PROJECT']}, auth_header=auth_header)
-    if len(project)>1:
-        print('Found several matching projects.')
-        return
-    if not project:
-        print("Project {} not found.".format(conf['STACKN_PROJECT']))
-        return False
+    project = get_projects(conf, params={'name': conf['STACKN_PROJECT']}, auth_header=auth_header)
+
+    if project == False:
+        return False, False, False
+    elif len(project)>1:
+        print('Found several matching projects. Please select a specific project.')
+        return False, False, False
+    elif len(project)==0:
+        print("Project \'{}\' not found.".format(conf['STACKN_PROJECT']))
+        return False, False, False
 
     project = project[0]
     endpoints = get_endpoints(conf['STACKN_URL'])
     url = endpoints[endpoint_type].format(project['id'])
     return conf, auth_header, url
+
+# Get functions
+
+def get_endpoints(studio_url):
+    # These endpoints are related to the API available in Studio under: stackn/components/studio/api/
+    endpoints = dict()
+    if (not 'http://' in studio_url) and (not 'https://' in studio_url):
+        studio_url = 'https://'+studio_url
+    base = studio_url.strip('/')+'/api'
+    endpoints['models'] = base+'/projects/{}/models/'
+    endpoints['resources'] = base+'/projects/{}/resources/'
+    endpoints['appinstances'] = base+'/projects/{}/appinstances/'
+    endpoints['members'] = base+'/projects/{}/members/'
+    endpoints['flavors'] = base+'/projects/{}/flavors/'
+    endpoints['environments'] = base+'/projects/{}/environments/'
+    endpoints['s3'] = base+'/projects/{}/s3/'
+    endpoints['mlflow'] = base+'/projects/{}/mlflow/'
+    endpoints['project_del'] = base+'/projects/{}'
+    endpoints['projects'] = base+'/projects/'
+    endpoints['project_templates'] = base+'/projecttemplates/'
+    endpoints['admin'] = dict()
+    endpoints['admin']['apps'] = base+'/apps/'
+    return endpoints
+
+def get_auth_header(conf):
+    conf, status = stackn.auth.get_token(conf)
+    if not status:
+        return False, False
+    auth_header = {"Authorization": "Token {}".format(conf['STACKN_ACCESS_TOKEN'])}
+    return auth_header, conf
+
+def get_current(secure):
+
+    res = {'STACKN_URL': False, 'STACKN_PROJECT': False}
+    conf, status = stackn.auth.get_config({'STACKN_SECURE': secure})
+
+    if not status:
+        print("Failed to get current STACKn configuration file.")
+    else:
+        if 'STACKN_URL' in conf:
+            res['STACKN_URL'] = conf['STACKN_URL']
+        if 'STACKN_PROJECT' in conf:
+            res['STACKN_PROJECT'] = conf['STACKN_PROJECT']
+    
+    return res
+
+def get_projects(conf={}, params=[], auth_header=[]):
+
+    conf, status = stackn.auth.get_config(conf)
+
+    if not status:
+        print('Cannot list projects.')
+        return False
+
+    auth_header, conf = get_auth_header(conf)
+
+    if not auth_header:
+        return False
+
+    endpoints = get_endpoints(conf['STACKN_URL'])
+    url = endpoints['projects']
+    if params:
+        r = requests.get(url, headers=auth_header, params=params, verify=conf['STACKN_SECURE'])
+    else:
+        r = requests.get(url, headers=auth_header, verify=conf['STACKN_SECURE'])
+    if r:
+        projects = json.loads(r.content)
+        return projects
+    else:
+        print("Fetching projects failed.")
+        print('Returned status code: {}'.format(r.status_code))
+        print('Reason: {}'.format(r.reason))
+        return None
+
+
+def get_remote(inp_conf):
+    
+    conf, status = stackn.auth.get_config(inp_conf)
+
+    if not status:
+        return False
+
+    keys = stackn.auth._get_remote(conf)
+    
+    if not keys:
+        return False
+
+    return keys
+
+
+# Create functions
 
 def create_template(template='template.json', image="image.png", studio_url=[], secure_mode=True):
     
@@ -242,6 +257,7 @@ def create_template(template='template.json', image="image.png", studio_url=[], 
         print(r.text)
         print(r.reason)
 
+
 def create_templates(folder='.', studio_url=[], secure_mode=True):
     subfolders = [f.path for f in os.scandir(folder) if f.is_dir()]
     curr_dir = os.getcwd()
@@ -250,6 +266,7 @@ def create_templates(folder='.', studio_url=[], secure_mode=True):
         create_template(studio_url=studio_url, secure_mode=secure_mode)
         os.chdir(curr_dir)
 
+
 def create_apps(folder='.', studio_url=[], secure_mode=True):
     subfolders = [f.path for f in os.scandir(folder) if f.is_dir()]
     curr_dir = os.getcwd()
@@ -257,6 +274,7 @@ def create_apps(folder='.', studio_url=[], secure_mode=True):
         os.chdir(folder)
         create_app(studio_url=studio_url, secure_mode=secure_mode)
         os.chdir(curr_dir)
+
 
 def create_app(settings="config.json",
                chart_archive="chart",
@@ -360,7 +378,8 @@ def create_project(name,
         print('Status code: {}'.format(res.status_code))
         print(res.text)
 
-def create_resource(filename, studio_url, project, secure):
+
+def create_meta_resource(filename, studio_url, project, secure):
     conf = {
         'STACKN_URL': studio_url,
         'STACKN_PROJECT': project,
@@ -400,6 +419,7 @@ def create_resource(filename, studio_url, project, secure):
         print('Failed to create resource.')
         print('Status code: {}'.format(res.status_code))
         print(res.text)
+
 
 def create_object(model_name,
                   studio_url=[],
@@ -507,39 +527,35 @@ def create_object(model_name,
 
     return True
 
-def create_releasename(name, studio_url, project, secure):
-    conf = {
-        'STACKN_URL': studio_url,
-        'STACKN_PROJECT': project,
-        'STACKN_SECURE': secure
-    }
-    conf, auth_header, url = setup_project_endpoint_call(conf, 'releasenames')
-    params = {
-        "name": name
-    }
-    res = requests.post(url, json=params, headers=auth_header, verify=conf['STACKN_SECURE'])
-    if res:
-        print(res.text)
-    else:
-        print("Failed to create release name.")
-        print('Status code: {}'.format(res.status_code))
-        print(res.text)
 
+# Delete functions
 
 def delete_app(name, studio_url=[], project=[], secure=True):
+    
     conf = {
         "STACKN_URL": studio_url,
         "STACKN_PROJECT": project,
         "STACKN_SECURE": secure
     }
+
     conf, auth_header, url = setup_project_endpoint_call(conf, 'appinstances')
+    
+    if not conf or not auth_header or not url:
+        print("Failed to set up project API endpoint call.")
+        return False
+
     params = {
         "name": name
     }
-    apps = call_project_endpoint('appinstances', params=params)
+
+    apps = call_project_endpoint('appinstances', conf, params=params)
+
+    if not apps:
+        return False
+    
     if len(apps) > 1:
         print("Found multiple apps with that name, deleting all...")
-    if len(apps) == 0:
+    elif len(apps) == 0:
         print("Found no app with that name, aborting...")
         return False
 
@@ -554,37 +570,6 @@ def delete_app(name, studio_url=[], project=[], secure=True):
             print('Status code: {}'.format(res.status_code))
             print(res.text)
 
-def delete_app_obj(slug, studio_url=[], secure=True):
-    conf = {
-        'STACKN_URL': studio_url,
-        'STACKN_SECURE': secure
-    }
-    conf, status = stackn.auth.get_config(conf, required=['STACKN_URL'])
-
-    if not status:
-        print("Failed to get current STACKn configuration file.")
-        return False
-
-    auth_header, conf = get_auth_header(conf)
-    if not auth_header:
-        return False
-
-    payload = {
-        'slug': slug
-    }
-    endpoints = get_endpoints(conf['STACKN_URL'])
-    url = endpoints['admin']['apps']
-
-    apps = requests.get(url, headers=auth_header, params=payload, verify=conf['STACKN_SECURE'])
-    apps = apps.json()
-    for app in apps:
-        print("Deleting {}, revision {}.".format(app['name'], app['revision']))
-        print(url+str(app['id'])+'/')
-        r = requests.delete(url+str(app['id'])+'/', headers=auth_header, verify=conf['STACKN_SECURE'])
-        print(r.text)
-
-    print(url)
-
 
 def delete_object(name, version=None, studio_url=[], project=[], secure=True):
     if version:
@@ -597,29 +582,47 @@ def delete_object(name, version=None, studio_url=[], project=[], secure=True):
         "STACKN_PROJECT": project,
         "STACKN_SECURE": secure
     }
+
     conf, auth_header, url = setup_project_endpoint_call(conf, 'models')
+
+    if not conf or not auth_header or not url:
+        print("Failed to setup project API endpoint.")
+        return False
+
     objects = call_project_endpoint('models', conf=conf, params=params)
 
-    if not objects:
-        print("No objects found with the given name and/or version.")
+    if objects == False:
+        return False
+    elif len(objects) == 0:
+        print("No model objects found with the given name and/or version.")
+
     for obj in objects:
         tmp_url = '{}{}/'.format(url, obj['id'])
         res = requests.delete(tmp_url, headers=auth_header, verify=conf['STACKN_SECURE'])
         if res:
-            print("Deleted object: {}".format(name))
+            print("Deleted model object: {}".format(name))
         else:
-            print("Failed to delete object.")
+            print("Failed to delete model object.")
             print('Status code: {}'.format(res.status_code))
             print(res.text)
 
+
 def delete_project(name, studio_url=[], secure=True):
+    
     conf = {
         "STACKN_URL": studio_url,
         "STACKN_SECURE": secure,
         "STACKN_PROJECT": name
     }
+    
     conf, auth_header, url = setup_project_endpoint_call(conf, 'project_del')
+
+    if not conf or not auth_header or not url:
+        print("Failed to set up project API endpoint")
+        return False
+    
     res = requests.delete(url, headers=auth_header, verify=conf['STACKN_SECURE'])
+    
     if res:
         print("Deleted project: {}".format(name))
     else:
@@ -627,26 +630,43 @@ def delete_project(name, studio_url=[], secure=True):
         print('Status code: {}'.format(res.status_code))
         print(res.text)
 
+
 def delete_meta_resource(resource_type, name, project=[], studio_url=[], secure=True):
+
     conf = {
         "STACKN_PROJECT": project,
         "STACKN_URL": studio_url,
         "STACKN_SECURE": secure
     }
+
     conf, auth_header, url = setup_project_endpoint_call(conf, resource_type)
+
+    if not conf or not auth_header or not url:
+        print("Failed to setup project API endpoint.")
+        return False
+    
     params = {
         "name": name
     }
-    envs = call_project_endpoint(resource_type, conf=conf, params=params)
-    if len(envs) == 0:
-        print("Didn't find environment with that name.")
+
+    prj_endpts = call_project_endpoint(resource_type, conf=conf, params=params)
+    
+    if prj_endpts == False:
         return False
-    elif len(envs) > 1:
-        print("Found multiple environments with that name.")
+    elif len(prj_endpts) == 0:
+        print("No {}: \'{}\' associated with the current project".format(resource_type, name))
+        return False
+    elif len(prj_endpts) > 1:
+        print("Found multiple resources with the passed name.")
         return False
     
-    env = envs[0]
-    url = '{}{}/'.format(url, env['id'])
+    endpt = prj_endpts[0]
+
+    if resource_type == "mlflow" or resource_type == "s3":
+        url = '{}{}/'.format(url, endpt['name'])
+    else:    
+        url = '{}{}/'.format(url, endpt['id'])
+
     res = requests.delete(url, headers=auth_header, verify=conf['STACKN_SECURE'])
     if res:
         print("Deleted {}: {}".format(resource_type, name))
@@ -655,6 +675,7 @@ def delete_meta_resource(resource_type, name, project=[], studio_url=[], secure=
         print('Status code: {}'.format(res.status_code))
         print(res.text)
 
+# Set Function
 
 def set_current(conf):
     res = stackn.auth._set_current(conf)
